requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
jinja2==3.1.2
groq==0.4.1
pgvector==0.2.4
numpy==1.24.3
pydantic==2.5.0